+++ test/functional/dd-functional-settlement-spec.js
@@ -83,7 +83,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     page.confirmVerification();
     expect( page.totalCostOfAttendance.getText() ).toEqual( '43,626' );
     expect( page.studentTotalCost.getText() ).toEqual( '32,026' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-474' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '2,526' );
   } );
 
   it( 'should properly update when the housing and meals are modified', function() {
@@ -92,7 +92,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     expect( page.totalCostOfAttendance.getText() ).toEqual( '42,626' );
     expect( page.studentTotalCost.getText() ).toEqual( '31,026' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-1,474' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '1,526' );
   } );
 
   it( 'should properly update when the transportation is modified', function() {
@@ -101,7 +101,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     expect( page.totalCostOfAttendance.getText() ).toEqual( '43,526' );
     expect( page.studentTotalCost.getText() ).toEqual( '31,926' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-574' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '2,426' );
   } );
 
   it( 'should properly update when the books and supplies are modified', function() {
@@ -110,7 +110,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     expect( page.totalCostOfAttendance.getText() ).toEqual( '43,726' );
     expect( page.studentTotalCost.getText() ).toEqual( '32,126' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-374' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '2,626' );
   } );
 
   it( 'should properly update when the other education costs are modified', function() {
@@ -119,7 +119,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     expect( page.totalCostOfAttendance.getText() ).toEqual( '44,126' );
     expect( page.studentTotalCost.getText() ).toEqual( '32,526' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '26' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '3,026' );
   } );
 
   //  Grants and scholarships onload: $7,600
@@ -132,7 +132,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     expect( page.totalGrantsScholarships.getText() ).toEqual( '15,600' );
     expect( page.studentTotalCost.getText() ).toEqual( '28,026' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-4,474' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '-1,474' );
   } );
 
   it( 'should properly update when the Federal Pell Grants are modified above the Federal limits', function() {
@@ -143,7 +143,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     // expect( EC.visibilityOf( page.pellGrantCapWarning ) );
     expect( page.totalGrantsScholarships.getText() ).toEqual( '15,915' );
     expect( page.studentTotalCost.getText() ).toEqual( '27,711' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-4,789' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '-1,789' );
   } );
 
   // TODO: Uncomment this once it's built in the JS code
@@ -198,7 +198,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     expect( page.totalGrantsScholarships.getText() ).toEqual( '12,500' );
     expect( page.studentTotalCost.getText() ).toEqual( '31,126' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-1,374' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '1,626' );
   } );
 
   // TODO: Uncomment this once it's built in the design, HTML, and JS code
@@ -237,25 +237,28 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     expect( page.remainingCostFinal.getText() ).toEqual( '38526' );
   } ); */
 
-  // Personal and family contributions onload: $14,000
-  // Total cost: $36,026
-  // Remaining cost: $7,526
-
   it( 'should properly update when the cash a student will personally provide is modified', function() {
     page.confirmVerification();
     page.setStudentContribution( 1500 );
     browser.sleep( 750 );
     expect( page.totalContributions.getText() ).toEqual( '19,500' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-1,974' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '1,026' );
   } );
 
-  // TODO: Figure out WHY input on this is broken now. It wasn't before!
   it( 'should properly update when the cash a student\'s family will provide is modified', function() {
     page.confirmVerification();
     page.setFamilyContribution( 4000 );
     browser.sleep( 750 );
     expect( page.totalContributions.getText() ).toEqual( '8,000' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '9,526' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '12,526' );
+  } );
+
+  it( 'should properly update when the Parent PLUS loan is modified', function() {
+    page.confirmVerification();
+    page.setParentPlusContribution( 2000 );
+    browser.sleep( 750 );
+    expect( page.totalContributions.getText() ).toEqual( '19,000' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '1,526' );
   } );
 
   it( 'should properly update when the work study earnings are modified within the allowed limit', function() {
@@ -263,7 +266,7 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     page.setWorkStudyContribution( 2000 );
     browser.sleep( 750 );
     expect( page.totalContributions.getText() ).toEqual( '17,000' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '526' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '3,526' );
   } );
 
   // TODO: Uncomment this once it's built in the JS code
@@ -293,40 +296,49 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
   // TODO: Uncomment ALL program debt and repayment once it's built in the JS
 
 
-  it( 'should properly update when the federal Perkins loans are modified within the allowed limit', function() {
-    page.confirmVerification();
-    page.setFederalPerkinsLoans( 2000 );
-    browser.sleep( 750 );
-    expect( page.totalFederalLoans.getText() ).toEqual( '7,500' );
-    expect( page.totalDebt.getText() ).toEqual( '13,500' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '526' );
-    // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
-    // expect( page.totalRepayment.getText() ).toEqual( '10000' );
-    // TODO: expect the est. monthly student loan expense is recalculated
-  } );
+  // The following tests could be useful for non-settlement, so I'm
+  // salvaging it for the future --mistergone
+  //
+  // it( 'should properly update when the federal Perkins loans are modified within the allowed limit', function() {
+  //   page.confirmVerification();
+  //   page.setFederalPerkinsLoans( 2000 );
+  //   browser.sleep( 750 );
+  //   expect( page.totalFederalLoans.getText() ).toEqual( '7,500' );
+  //   expect( page.totalDebt.getText() ).toEqual( '13,500' );
+  //   expect( page.remainingCostFinal.getText() ).toEqual( '526' );
+  //   // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
+  //   // expect( page.totalRepayment.getText() ).toEqual( '10000' );
+  //   // TODO: expect the est. monthly student loan expense is recalculated
+  // } );
 
-  it( 'should properly update when the federal Perkins loans are modified above the allowed limit', function() {
+  // it( 'should properly update when the federal Perkins loans are modified above the allowed limit', function() {
+  //   page.confirmVerification();
+  //   page.setFederalPerkinsLoans( 15000 );
+  //   browser.sleep( 750 );
+  //   // TODO: expect student is informed about the Perkins loan cap
+  //   // expect( EC.visibilityOf( page.perkinsLoanCapWarning ) );
+  //   // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
+  //   expect( page.totalFederalLoans.getText() ).toEqual( '11,000' );
+  //   expect( page.totalDebt.getText() ).toEqual( '17,000' );
+  //   expect( page.remainingCostFinal.getText() ).toEqual( '-2,974' );
+  //   // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
+  //   // expect( page.totalRepayment.getText() ).toEqual( '10000' );
+  //   // TODO: expect the est. monthly student loan expense is recalculated
+  // } );
+
+  it( 'should properly hide the federal Perkins loans when the school does not offer it', function() {
     page.confirmVerification();
-    page.setFederalPerkinsLoans( 15000 );
     browser.sleep( 750 );
-    // TODO: expect student is informed about the Perkins loan cap
-    // expect( EC.visibilityOf( page.perkinsLoanCapWarning ) );
-    // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
-    expect( page.totalFederalLoans.getText() ).toEqual( '11,000' );
-    expect( page.totalDebt.getText() ).toEqual( '17,000' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-2,974' );
-    // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
-    // expect( page.totalRepayment.getText() ).toEqual( '10000' );
-    // TODO: expect the est. monthly student loan expense is recalculated
+    expect( page.federalPerkinsLoans.isDisplayed() ).toBeFalsy();
   } );
 
   it( 'should properly update when the federal subsidized loans are modified within the allowed limit', function() {
     page.confirmVerification();
     page.setSubsidizedLoans( 3000 );
     browser.sleep( 750 );
-    expect( page.totalFederalLoans.getText() ).toEqual( '8,000' );
-    expect( page.totalDebt.getText() ).toEqual( '14,000' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '26' );
+    expect( page.totalFederalLoans.getText() ).toEqual( '5,000' );
+    expect( page.totalDebt.getText() ).toEqual( '11,000' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '3,026' );
     // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
     // expect( page.totalRepayment.getText() ).toEqual( '10000' );
     // TODO: expect the est. monthly student loan expense is recalculated
@@ -338,9 +350,9 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     // TODO: expect student is informed about the subsidized loan cap
     // expect( EC.visibilityOf( page.subsidizedLoanCapWarning ) );
-    expect( page.totalFederalLoans.getText() ).toEqual( '8,500' );
-    expect( page.totalDebt.getText() ).toEqual( '14,500' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-474' );
+    expect( page.totalFederalLoans.getText() ).toEqual( '5,500' );
+    expect( page.totalDebt.getText() ).toEqual( '11,500' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '2,526' );
     // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
     // expect( page.totalRepayment.getText() ).toEqual( '10000' );
     // TODO: expect the est. monthly student loan expense is recalculated
@@ -350,9 +362,9 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     page.confirmVerification();
     page.setUnsubsidizedLoans( 3000 );
     browser.sleep( 750 );
-    expect( page.totalFederalLoans.getText() ).toEqual( '9,500' );
-    expect( page.totalDebt.getText() ).toEqual( '15,500' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-1,474' );
+    expect( page.totalFederalLoans.getText() ).toEqual( '6,500' );
+    expect( page.totalDebt.getText() ).toEqual( '12,500' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '1,526' );
     // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
     // expect( page.totalRepayment.getText() ).toEqual( '10000' );
     // TODO: expect the est. monthly student loan expense is recalculated
@@ -364,9 +376,9 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     // TODO: expect student is informed about the unsubsidized loan cap
     // expect( EC.visibilityOf( page.unsussidizedLoanCapWarning ) );
-    expect( page.totalFederalLoans.getText() ).toEqual( '12,500' );
-    expect( page.totalDebt.getText() ).toEqual( '18,500' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-4,474' );
+    expect( page.totalFederalLoans.getText() ).toEqual( '9,500' );
+    expect( page.totalDebt.getText() ).toEqual( '15,500' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '-1,474' );
     // expect( page.totalProgramDebt.getText() ).toEqual( '3000' );
     // expect( page.totalRepayment.getText() ).toEqual( '10000' );
     // TODO: expect the est. monthly student loan expense is recalculated
@@ -432,8 +444,8 @@ fdescribe( 'A dynamic financial aid disclosure that\'s required by settlement', 
     browser.sleep( 750 );
     expect( page.privateLoanInterestRate.getAttribute('value') ).toBeGreaterThan( 0 );
     expect( page.totalPrivateLoansPaymentPlans.getText() ).toEqual( '7,000' );
-    expect( page.totalDebt.getText() ).toEqual( '15,500' );
-    expect( page.remainingCostFinal.getText() ).toEqual( '-1,474' );
+    expect( page.totalDebt.getText() ).toEqual( '12,500' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '1,526' );
     // expect( page.totalProgramDebt.getText() ).toEqual( '?' );
     // expect( page.totalRepayment.getText() ).toEqual( '?' );
     // TODO: expect the est. monthly student loan expense is recalculated
@@ -485,8 +497,8 @@ it( 'should properly update when more than one private loans is modified', funct
   it( 'should display proper debt values', function() {
     browser.sleep( 1000 );
     page.confirmVerification();
-    expect( page.totalProgramDebt.getText() ).toEqual( '29,000' );
-    expect( page.totalRepayment.getText() ).toEqual( '37,525' );
+    expect( page.totalProgramDebt.getText() ).toEqual( '23,000' );
+    expect( page.totalRepayment.getText() ).toEqual( '29,889' );
   } );
 
   it( 'should update total borrowing and verbiage when program length is changed', function() {
@@ -494,7 +506,7 @@ it( 'should properly update when more than one private loans is modified', funct
      page.setProgramLength( 4 );
      browser.sleep( 1000 );
      expect( page.futureYearsAttending.getText() ).toEqual( 'four' );
-     expect( page.totalProgramDebt.getText() ).toEqual( '58,000' );
+     expect( page.totalProgramDebt.getText() ).toEqual( '46,000' );
   });
 
   it( 'should properly describe a future based on not covering enough of the cost of college that is needed', function() {
@@ -504,35 +516,36 @@ it( 'should properly update when more than one private loans is modified', funct
     browser.sleep( 750 );
     expect( page.futurePositiveRemainingCost.isDisplayed() ).toBeTruthy();
     expect( page.futureNegativeRemainingCost.isDisplayed() ).toBeFalsy();
-    expect( page.remainingCostFinal.getText() ).toEqual( '3,526' );
-    expect( page.futureTotalLoans.getText() ).toEqual( '$14,500' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '6,526' );
+    expect( page.futureTotalLoans.getText() ).toEqual( '$11,500' );
     expect( page.futureYearsAttending.getText() ).toEqual( 'two' );
-    expect( page.futureTotalDebt.getText() ).toEqual( '$37,525' );
+    expect( page.futureTotalDebt.getText() ).toEqual( '$29,889' );
   } );
 
   it( 'should properly describe a future based on covering more of the cost of college that is needed', function() {
     browser.sleep( 1000 );
     page.confirmVerification();
+    page.setFamilyContribution( 25000 );
     browser.sleep( 750 );
     expect( page.futurePositiveRemainingCost.isDisplayed() ).toBeFalsy();
     expect( page.futureNegativeRemainingCost.isDisplayed() ).toBeTruthy();
-    expect( page.remainingCostFinal.getText() ).toEqual( '-474' );
-    expect( page.futureTotalLoans.getText() ).toEqual( '$14,500' );
+    expect( page.remainingCostFinal.getText() ).toEqual( '-8,474' );
+    expect( page.futureTotalLoans.getText() ).toEqual( '$11,500' );
     expect( page.futureYearsAttending.getText() ).toEqual( 'two' );
-    expect( page.futureTotalDebt.getText() ).toEqual( '$37,525' );
+    expect( page.futureTotalDebt.getText() ).toEqual( '$29,889' );
   } );
 
   it( 'should properly describe a future based on covering exactly the cost of college that is needed', function() {
     browser.sleep( 1000 );
     page.confirmVerification();
-    page.setFamilyContribution( 13526 );
+    page.setFamilyContribution( 16526 );
     browser.sleep( 750 );
     expect( page.futurePositiveRemainingCost.isDisplayed() ).toBeFalsy();
     expect( page.futureNegativeRemainingCost.isDisplayed() ).toBeFalsy();
     expect( page.remainingCostFinal.getText() ).toEqual( '0' );
-    expect( page.futureTotalLoans.getText() ).toEqual( '$14,500' );
+    expect( page.futureTotalLoans.getText() ).toEqual( '$11,500' );
     expect( page.futureYearsAttending.getText() ).toEqual( 'two' );
-    expect( page.futureTotalDebt.getText() ).toEqual( '$37,525' );
+    expect( page.futureTotalDebt.getText() ).toEqual( '$29,889' );
   } );
 
   it( 'should let a student continue on to Step 2', function() {